Support for toggling overrides at runtime through presence of a file.
If CHECK_FILE is defined as a filename, it is checked any time *sync is
called. If the file is not present, all calls are passed unmodified
through to libc.

This allows libeatmydata to be turned on systemwide in /etc/ld.so.preload
and only activated when a specific file such as "/eatmydata" is present.
Handy for scripting laptop-mode to avoid spinups, temporary latency
reduction when playing music with slow SSDs, etc.